Fix styling to match desired outcome
While I mostly got this, I did have difficulty with the center alignment
of the button container text. I thought I would use auto padding on the
left and right to center, but that didn't work. Text-align was the
obvious answer; I was too focused on the immediate lesson I just saw!