Add flex styling to header
This was tricky. The solution I originally came up with worked, using
the classes for each side of the header and styling them separately, but
after viewing the solution I saw how much more concise the code could be
by only using a few lines in the header and the gap attribute in ul.