Remove redundant styling
After reviewing the solution, I saw several places for making the code
more concise. I removed all the flex styling from the plantblock divs
and just used the max-width for 200px. Didn't need the flex at all since
the img element is already block style and so is above the text by
default. Used a body style to center the text, and then could remove the
centering of the title and the text class divs. Still so much to learn!