Add divs to group items together for flexbox
Added a .columns div to separate the exclamation point on the left from
the everything else on the right. Then added a .top div to enclose the
header and close buttons. The text box was already its own div. Finally
added a .bottom div for the continue and cancel buttons.